1-strdup.c

+++ 0x0B-malloc_free/1-strdup.c
@@ -12,8 +12,7 @@
 char *_strdup(char *str)
 {
 	char *dest;
-	long unsigned int i;
-	
+	unsigned int i;
 
 	if (str == NULL)
 	{
